profesores no puede borrar ni editar

+++ 2º Trimestre/proyecto/delete.php
@@ -1,23 +1,44 @@
 <?php include "header.php" ?>
-<?php 
-    if(isset($_GET['eliminar']))
-    {
-        $id= htmlspecialchars($_GET['eliminar']);
 
-        $user = $_SESSION['usuario'];
+<?php
 
-        
 
-        if($_SESSION['admin']!="Administrador"){
-            $query = "DELETE FROM incidencias WHERE id = {$id} AND user='$user'"; 
-        }else{
-            $query = "DELETE FROM incidencias WHERE id = {$id}"; 
-        }
+
+?>
+
+<?php
+
+    session_start();
+    session_set_cookie_params(120);
+
+    if($_SESSION['admin']=="Profesor"){
+        echo "<script> alert ('No tienes suficientes permisos para acceder a esta página') 
+        window.location='http://pepe.thsite.top/proyecto/incidencias.php'</script>";
+    } else {
+
+        if(isset($_GET['eliminar']))
+        {
+            $id= htmlspecialchars($_GET['eliminar']);
     
+            $user = $_SESSION['usuario'];
+    
+            
+    
+            if($_SESSION['admin']!="Administrador"){
+                $query = "DELETE FROM incidencias WHERE id = {$id} AND user='$user'"; 
+            }else{
+                $query = "DELETE FROM incidencias WHERE id = {$id}"; 
+            }
+        
+    
+            $delete_query= $conn->query($query);
+            // header("Location: home.php");
+            echo "<script>window.location='incidencias.php';</script>";
+        }
 
-        $delete_query= $conn->query($query);
-        // header("Location: home.php");
-        echo "<script>window.location='incidencias.php';</script>";
     }
+
+
+
 ?>
 <?php include "footer.php" ?>
